Single source of truth for CSharp Parser version
