+++ src/dotnet/APIView/APIView/Languages/CodeFileBuilder.cs
@@ -1,17 +1,19 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using APIView;
-using APIView.Analysis;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.SymbolDisplay;
+using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.ComponentModel;
 using System.Linq;
+using ApiView;
+using APIView.Analysis;
+using APIView.Model.V2;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.SymbolDisplay;
 
-namespace ApiView
+namespace APIView.Languages
 {
     public class CodeFileBuilder
     {
@@ -76,153 +78,167 @@ namespace ApiView
             {
                 analyzer.VisitAssembly(assemblySymbol);
             }
-            var builder = new CodeFileTokensBuilder();
 
-            BuildDependencies(builder, dependencies);
-            BuildInternalsVisibleToAttributes(builder, assemblySymbol);
+            var codeFile = new CodeFile()
+            {
+                Language = "C#",
+                ParserVersion = CurrentVersion,
+                PackageName = assemblySymbol.Name,
+                PackageVersion = GetPackageVersion(assemblySymbol) ?? assemblySymbol.Identity.Version.ToString()
+            };
 
-            var navigationItems = new List<NavigationItem>();
+            if (dependencies != null)
+            {
+                BuildDependencies(codeFile.ReviewLines, dependencies);
+            }
+            BuildInternalsVisibleToAttributes(codeFile.ReviewLines, assemblySymbol);
+
             foreach (var namespaceSymbol in SymbolOrderProvider.OrderNamespaces(EnumerateNamespaces(assemblySymbol)))
             {
                 if (namespaceSymbol.IsGlobalNamespace)
                 {
                     foreach (var namedTypeSymbol in SymbolOrderProvider.OrderTypes(namespaceSymbol.GetTypeMembers()))
                     {
-                        BuildType(builder, namedTypeSymbol, navigationItems, false);
+                        BuildType(codeFile.ReviewLines, namedTypeSymbol, false);
                     }
                 }
                 else
                 {
-                    BuildNamespace(builder, namespaceSymbol, navigationItems);
+                    BuildNamespace(codeFile.ReviewLines, namespaceSymbol);
                 }
             }
 
-            NavigationItem assemblyNavigationItem = new NavigationItem()
-            {
-                Text = assemblySymbol.Name + ".dll",
-                ChildItems = navigationItems.ToArray(),
-                Tags = { { "TypeKind", "assembly" } }
-            };
-
-            var node = new CodeFile()
-            {
-                Name = $"{assemblySymbol.Name} ({assemblySymbol.Identity.Version})",
-                Language = "C#",
-                Tokens = builder.Tokens.ToArray(),
-                VersionString = CurrentVersion,
-                Navigation = new[] { assemblyNavigationItem },
-                Diagnostics = analyzer.Results.ToArray(),
-                PackageName = assemblySymbol.Name,
-                PackageVersion = assemblySymbol.Identity.Version.ToString()
-            };
-
-            return node;
+            codeFile.Diagnostics = analyzer.Results.ToArray();
+            VerifyCodeFile(codeFile);
+            return codeFile;
         }
 
-        public static void BuildInternalsVisibleToAttributes(CodeFileTokensBuilder builder, IAssemblySymbol assemblySymbol)
+        public static void BuildInternalsVisibleToAttributes(List<ReviewLine> reviewLines, IAssemblySymbol assemblySymbol)
         {
             var assemblyAttributes = assemblySymbol.GetAttributes()
                 .Where(a =>
-                    a.AttributeClass.Name == "InternalsVisibleToAttribute" &&
-                    !a.ConstructorArguments[0].Value.ToString().Contains(".Tests") &&
-                    !a.ConstructorArguments[0].Value.ToString().Contains(".Perf") &&
-                    !a.ConstructorArguments[0].Value.ToString().Contains("DynamicProxyGenAssembly2"));
+                    a.AttributeClass?.Name == "InternalsVisibleToAttribute" &&
+                    !a.ConstructorArguments[0].Value?.ToString()?.Contains(".Tests") == true &&
+                    !a.ConstructorArguments[0].Value?.ToString()?.Contains(".Perf") == true &&
+                    !a.ConstructorArguments[0].Value?.ToString()?.Contains("DynamicProxyGenAssembly2") == true);
             if (assemblyAttributes != null && assemblyAttributes.Any())
             {
-                builder.Append("Exposes internals to:", CodeFileTokenKind.Text);
-                builder.NewLine();
+                var internalVisibleLine = new ReviewLine()
+                {
+                    LineId = "InternalsVisibleTo",
+                    Tokens = [
+                        ReviewToken.CreateStringLiteralToken("Exposes internals to:")
+                    ]
+                };
+                reviewLines.Add(internalVisibleLine);
+
                 foreach (AttributeData attribute in assemblyAttributes)
                 {
                     if (attribute.ConstructorArguments.Length > 0)
                     {
-                        var param = attribute.ConstructorArguments[0].Value.ToString();
-                        var firstComma = param.IndexOf(',');
-                        param = firstComma > 0 ? param[..firstComma] : param;
-                        builder.Append(new CodeFileToken(param, CodeFileTokenKind.Text)
+                        var param = attribute.ConstructorArguments[0].Value?.ToString();
+                        if (!String.IsNullOrEmpty(param))
                         {
-                            // allow assembly to have a comment
-                            DefinitionId = attribute.AttributeClass.Name
-                        });
+                            var firstComma = param?.IndexOf(',');
+                            param = firstComma > 0 ? param?[..(int)firstComma] + "_" + param?[^5..] : param;
+                            reviewLines.Add(new ReviewLine()
+                            {
+                                LineId = attribute.AttributeClass?.Name + "_" + param,
+                                Tokens = [
+                                    ReviewToken.CreateStringLiteralToken(param)
+                                ]
+                            });
+                        }
                     }
-                    builder.NewLine();
                 }
-                builder.NewLine();
+                // Add an empty line after internals visible to section
+                reviewLines.Add(new ReviewLine() { RelatedToLine = internalVisibleLine.LineId });
             }
         }
 
-        public static void BuildDependencies(CodeFileTokensBuilder builder, List<DependencyInfo> dependencies)
+        public static void BuildDependencies(List<ReviewLine> reviewLines, List<DependencyInfo> dependencies)
         {
             if (dependencies != null && dependencies.Any())
             {
-                builder.NewLine();
-                builder.Append("Dependencies:", CodeFileTokenKind.Text);
-                builder.NewLine();
+                //Dependencies
+                var headerLine = new ReviewLine()
+                {
+                    LineId = "Dependencies"
+                };
+                var depToken = ReviewToken.CreateStringLiteralToken("Dependencies:");
+                depToken.NavigationDisplayName = "Dependencies";
+                depToken.RenderClasses.Add("dependencies");
+                headerLine.Tokens.Add(depToken);
+                reviewLines.Add(headerLine);
+
                 foreach (DependencyInfo dependency in dependencies)
                 {
-                    builder.Append(new CodeFileToken(dependency.Name, CodeFileTokenKind.Text)
+                    var versionToken = ReviewToken.CreateStringLiteralToken($"-{dependency.Version}");
+                    versionToken.SkipDiff = true;
+                    var dependencyLine = new ReviewLine()
                     {
-                        // allow dependency to be commentable
-                        DefinitionId = dependency.Name
-                    });
-                    // don't include the version in the API sign-off diffs
-                    builder.Append(null, CodeFileTokenKind.SkipDiffRangeStart);
-                    builder.Append($"-{dependency.Version}", CodeFileTokenKind.Text);
-                    builder.Append(null, CodeFileTokenKind.SkipDiffRangeEnd);
-                    builder.NewLine();
+                        LineId = dependency.Name + versionToken.Value,
+                        Tokens = [
+                            ReviewToken.CreateStringLiteralToken(dependency.Name, false),
+                            versionToken
+                        ]
+                    };
+                    reviewLines.Add(dependencyLine);
                 }
-
-                builder.NewLine();
+                reviewLines.Add(new ReviewLine() { RelatedToLine = headerLine.LineId });
             }
         }
 
-        private void BuildNamespace(CodeFileTokensBuilder builder, INamespaceSymbol namespaceSymbol, List<NavigationItem> navigationItems)
+        private void BuildNamespace(List<ReviewLine> reviewLines, INamespaceSymbol namespaceSymbol)
         {
             bool isHidden = HasOnlyHiddenTypes(namespaceSymbol);
-
-            if (isHidden)
+            var namespaceLine = new ReviewLine()
             {
-                builder.Append(null, CodeFileTokenKind.HiddenApiRangeStart);
-            }
-            builder.Keyword(SyntaxKind.NamespaceKeyword);
-            builder.Space();
-            BuildNamespaceName(builder, namespaceSymbol);
-
-            builder.Space();
-            builder.Punctuation(SyntaxKind.OpenBraceToken);
-            builder.IncrementIndent();
-            builder.NewLine();
-
-            List<NavigationItem> namespaceItems = new List<NavigationItem>();
-            foreach (var namedTypeSymbol in SymbolOrderProvider.OrderTypes(namespaceSymbol.GetTypeMembers()))
-            {
-                BuildType(builder, namedTypeSymbol, namespaceItems, isHidden);
-            }
-
-            CloseBrace(builder);
-
-            var namespaceItem = new NavigationItem()
-            {
-                NavigationId = namespaceSymbol.GetId(),
-                Text = namespaceSymbol.ToDisplayString(),
-                ChildItems = namespaceItems.ToArray(),
-                Tags = { { "TypeKind", "namespace" } },
-                IsHiddenApi = isHidden
+                LineId = namespaceSymbol.GetId(),
+                Tokens = [
+                    ReviewToken.CreateKeywordToken("namespace")
+                ],
+                IsHidden = isHidden
             };
-            navigationItems.Add(namespaceItem);
-            if (isHidden)
+
+            BuildNamespaceName(namespaceLine, namespaceSymbol);
+            var nameSpaceToken = namespaceLine.Tokens.LastOrDefault();
+            if (nameSpaceToken != null)
             {
-                builder.Append(null, CodeFileTokenKind.HiddenApiRangeEnd);
+                nameSpaceToken.RenderClasses.Add("namespace");
+                nameSpaceToken.NavigationDisplayName = namespaceSymbol.ToDisplayString();
             }
+            namespaceLine.Tokens.Last().HasSuffixSpace = true;
+            namespaceLine.Tokens.Add(ReviewToken.CreatePunctuationToken("{"));
+
+            // Add each members in the namespace
+            foreach (var namedTypeSymbol in SymbolOrderProvider.OrderTypes(namespaceSymbol.GetTypeMembers()).OrderBy(s => s.GetId()))
+            {
+                BuildType(namespaceLine.Children, namedTypeSymbol, isHidden);
+            }
+
+            reviewLines.Add(namespaceLine);
+            reviewLines.Add(new ReviewLine()
+            {
+                Tokens = [
+                    ReviewToken.CreateStringLiteralToken("}")
+                ],
+                IsHidden = isHidden,
+                IsContextEndLine = true
+            });
+            //Add an empty line in the review after current name space.
+            reviewLines.Add(new ReviewLine() { IsHidden = isHidden, RelatedToLine = namespaceLine.LineId });
         }
 
-        private void BuildNamespaceName(CodeFileTokensBuilder builder, INamespaceSymbol namespaceSymbol)
+        private void BuildNamespaceName(ReviewLine namespaceLine, INamespaceSymbol namespaceSymbol)
         {
             if (!namespaceSymbol.ContainingNamespace.IsGlobalNamespace)
             {
-                BuildNamespaceName(builder, namespaceSymbol.ContainingNamespace);
-                builder.Punctuation(SyntaxKind.DotToken);
+                BuildNamespaceName(namespaceLine, namespaceSymbol.ContainingNamespace);
+                var punctuation = ReviewToken.CreatePunctuationToken(".", false);
+                namespaceLine.Tokens.Add(punctuation);
             }
-            NodeFromSymbol(builder, namespaceSymbol);
+            DisplayName(namespaceLine, namespaceSymbol, namespaceSymbol);
         }
 
         private bool HasAnyPublicTypes(INamespaceSymbol subNamespaceSymbol)
@@ -232,85 +248,96 @@ namespace ApiView
 
         private bool HasOnlyHiddenTypes(INamespaceSymbol namespaceSymbol)
         {
-            return namespaceSymbol.GetTypeMembers().All(t=> IsHiddenFromIntellisense(t) || !IsAccessible(t));
+            return namespaceSymbol.GetTypeMembers().All(t => IsHiddenFromIntellisense(t) || !IsAccessible(t));
         }
 
-        private void BuildType(CodeFileTokensBuilder builder, INamedTypeSymbol namedType, List<NavigationItem> navigationBuilder, bool inHiddenScope)
+        private void BuildType(List<ReviewLine> reviewLines, INamedTypeSymbol namedType, bool inHiddenScope)
         {
             if (!IsAccessible(namedType))
             {
                 return;
             }
 
-            bool isHidden = IsHiddenFromIntellisense(namedType);
-            var navigationItem = new NavigationItem()
+            bool isHidden = IsHiddenFromIntellisense(namedType) || inHiddenScope;
+            var reviewLine = new ReviewLine()
             {
-                NavigationId = namedType.GetId(),
-                Text = namedType.ToDisplayString(SymbolDisplayFormat.MinimallyQualifiedFormat),
-                IsHiddenApi = isHidden
+                LineId = namedType.GetId(),
+                IsHidden = isHidden
             };
-            navigationBuilder.Add(navigationItem);
-            navigationItem.Tags.Add("TypeKind", namedType.TypeKind.ToString().ToLowerInvariant());
 
-            if (isHidden && !inHiddenScope)
-            {
-                builder.Append(null, CodeFileTokenKind.HiddenApiRangeStart);
-            }
-
-            BuildDocumentation(builder, namedType);
-            BuildAttributes(builder, namedType.GetAttributes());
-
-            builder.WriteIndent();
-            BuildVisibility(builder, namedType);
-
-            builder.Space();
+            // Build documentation, attributes, visibility, and name
+            BuildDocumentation(reviewLines, namedType, isHidden, namedType.GetId());
+            BuildAttributes(reviewLines, namedType.GetAttributes(), isHidden, namedType.GetId());
+            BuildVisibility(reviewLine.Tokens, namedType);
 
             switch (namedType.TypeKind)
             {
                 case TypeKind.Class:
-                    BuildClassModifiers(builder, namedType);
-                    builder.Keyword(SyntaxKind.ClassKeyword);
+                    BuildClassModifiers(reviewLine.Tokens, namedType);
+                    reviewLine.Tokens.Add(ReviewToken.CreateKeywordToken(SyntaxKind.ClassKeyword));
                     break;
                 case TypeKind.Delegate:
-                    builder.Keyword(SyntaxKind.DelegateKeyword);
+                    reviewLine.Tokens.Add(ReviewToken.CreateKeywordToken(SyntaxKind.DelegateKeyword));
                     break;
                 case TypeKind.Enum:
-                    builder.Keyword(SyntaxKind.EnumKeyword);
+                    reviewLine.Tokens.Add(ReviewToken.CreateKeywordToken(SyntaxKind.EnumKeyword));
                     break;
                 case TypeKind.Interface:
-                    builder.Keyword(SyntaxKind.InterfaceKeyword);
+                    reviewLine.Tokens.Add(ReviewToken.CreateKeywordToken(SyntaxKind.InterfaceKeyword));
                     break;
                 case TypeKind.Struct:
                     if (namedType.IsReadOnly)
                     {
-                        builder.Keyword(SyntaxKind.ReadOnlyKeyword);
-                        builder.Space();
+                        reviewLine.Tokens.Add(ReviewToken.CreateKeywordToken(SyntaxKind.ReadOnlyKeyword));
                     }
-                    builder.Keyword(SyntaxKind.StructKeyword);
+                    reviewLine.Tokens.Add(ReviewToken.CreateKeywordToken(SyntaxKind.StructKeyword));
                     break;
             }
 
-            builder.Space();
+            DisplayName(reviewLine, namedType, namedType);
 
-            NodeFromSymbol(builder, namedType);
+            // Add navigation short name and render classes to Type name token. Navigation tree is built dynamically based on these properties
+            var typeToken = reviewLine.Tokens.FirstOrDefault(t => t.Kind == TokenKind.TypeName && string.IsNullOrEmpty(t.NavigateToId));
+            if (typeToken != null)
+            {
+                typeToken.NavigationDisplayName = namedType.ToDisplayString(SymbolDisplayFormat.MinimallyQualifiedFormat);
+                typeToken.RenderClasses.Add(namedType.TypeKind.ToString().ToLowerInvariant());
+            }
             if (namedType.TypeKind == TypeKind.Delegate)
             {
-                builder.Punctuation(SyntaxKind.SemicolonToken);
-                builder.NewLine();
+                reviewLine.Tokens.Last().HasSuffixSpace = false;
+                reviewLine.Tokens.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.SemicolonToken));
+                reviewLines.Add(reviewLine);
                 return;
             }
 
-            builder.Space();
-
-            BuildBaseType(builder, namedType);
-
-            builder.Punctuation(SyntaxKind.OpenBraceToken);
-            builder.IncrementIndent();
-            builder.NewLine();
+            reviewLine.Tokens.Last().HasSuffixSpace = true;
+            BuildBaseType(reviewLine, namedType);
+            reviewLine.Tokens.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.OpenBraceToken));
 
             foreach (var namedTypeSymbol in SymbolOrderProvider.OrderTypes(namedType.GetTypeMembers()))
             {
-                BuildType(builder, namedTypeSymbol, navigationBuilder, inHiddenScope || isHidden);
+                try
+                {
+                    // Check if this is an extension member container
+                    if (IsExtensionMemberContainer(namedTypeSymbol))
+                    {
+                        BuildExtensionMember(reviewLine.Children, namedTypeSymbol, isHidden);
+                    }
+                    else
+                    {
+                        BuildType(reviewLine.Children, namedTypeSymbol, isHidden);
+                    }
+                }
+                catch (Exception ex)
+                {
+                    // Log the exception before falling back to regular type building
+                    System.Console.Error.WriteLine($"Exception in extension member detection for type '{namedTypeSymbol?.Name}': {ex}");
+                    if (namedTypeSymbol != null)
+                    {
+                        BuildType(reviewLine.Children, namedTypeSymbol, isHidden);
+                    }
+                }
             }
 
             foreach (var member in SymbolOrderProvider.OrderMembers(namedType.GetMembers()))
@@ -327,68 +354,73 @@ namespace ApiView
                         continue;
                     }
                 }
-
-                BuildMember(builder, member, inHiddenScope);
+                BuildMember(reviewLine.Children, member, isHidden);
             }
-
-            CloseBrace(builder);
-
-            if (isHidden && !inHiddenScope)
+            reviewLines.Add(reviewLine);
+            reviewLines.Add(new ReviewLine()
             {
-                builder.Append(null, CodeFileTokenKind.HiddenApiRangeEnd);
-            }
+                Tokens = [
+                    ReviewToken.CreateStringLiteralToken("}")
+                ],
+                IsHidden = isHidden,
+                IsContextEndLine = true
+            });
+            reviewLines.Add(new ReviewLine() { IsHidden = isHidden, RelatedToLine = reviewLine.LineId });
         }
 
-        private void BuildDocumentation(CodeFileTokensBuilder builder, ISymbol symbol)
+        private void BuildDocumentation(List<ReviewLine> reviewLines, ISymbol symbol, bool isHidden, string relatedTo)
         {
-            var lines = symbol.GetDocumentationCommentXml().Trim().Split(_newlineChars);
-            if (lines.All(string.IsNullOrWhiteSpace))
+            var lines = symbol.GetDocumentationCommentXml()?.Trim().Split(_newlineChars);
+
+            if (lines != null)
             {
-                return;
+                if (lines.All(string.IsNullOrWhiteSpace))
+                {
+                    return;
+                }
+                foreach (var line in lines)
+                {
+                    var docToken = ReviewToken.CreateCommentToken("// " + line.Trim());
+                    docToken.IsDocumentation = true;
+                    reviewLines.Add(new ReviewLine()
+                    {
+                        Tokens = [docToken],
+                        IsHidden = isHidden,
+                        RelatedToLine = relatedTo
+                    });
+                }
             }
-            builder.Append(null, CodeFileTokenKind.DocumentRangeStart);
-            foreach (var line in lines)
-            {
-                builder.WriteIndent();
-                builder.Comment("// " + line.Trim());
-                builder.NewLine();
-            }
-            builder.Append(null, CodeFileTokenKind.DocumentRangeEnd);
         }
 
-        private static void BuildClassModifiers(CodeFileTokensBuilder builder, INamedTypeSymbol namedType)
+        private static void BuildClassModifiers(List<ReviewToken> tokenList, INamedTypeSymbol namedType)
         {
             if (namedType.IsAbstract)
             {
-                builder.Keyword(SyntaxKind.AbstractKeyword);
-                builder.Space();
+                tokenList.Add(ReviewToken.CreateKeywordToken(SyntaxKind.AbstractKeyword));
             }
 
             if (namedType.IsStatic)
             {
-                builder.Keyword(SyntaxKind.StaticKeyword);
-                builder.Space();
+                tokenList.Add(ReviewToken.CreateKeywordToken(SyntaxKind.StaticKeyword));
             }
 
             if (namedType.IsSealed)
             {
-                builder.Keyword(SyntaxKind.SealedKeyword);
-                builder.Space();
+                tokenList.Add(ReviewToken.CreateKeywordToken(SyntaxKind.SealedKeyword));
             }
         }
 
-        private void BuildBaseType(CodeFileTokensBuilder builder, INamedTypeSymbol namedType)
+        private void BuildBaseType(ReviewLine reviewLine, INamedTypeSymbol namedType)
         {
             bool first = true;
 
             if (namedType.BaseType != null &&
                 namedType.BaseType.SpecialType == SpecialType.None)
             {
-                builder.Punctuation(SyntaxKind.ColonToken);
-                builder.Space();
+                reviewLine.AddToken(ReviewToken.CreatePunctuationToken(SyntaxKind.ColonToken));
                 first = false;
-
-                DisplayName(builder, namedType.BaseType);
+                DisplayName(reviewLine, namedType.BaseType);
+                reviewLine.Tokens.Last().HasSuffixSpace = true;
             }
 
             foreach (var typeInterface in namedType.Interfaces)
@@ -397,131 +429,277 @@ namespace ApiView
 
                 if (!first)
                 {
-                    builder.Punctuation(SyntaxKind.CommaToken);
-                    builder.Space();
+                    reviewLine.Tokens.Last().HasSuffixSpace = false;
+                    reviewLine.AddToken(ReviewToken.CreatePunctuationToken(SyntaxKind.CommaToken));
                 }
                 else
                 {
-                    builder.Punctuation(SyntaxKind.ColonToken);
-                    builder.Space();
+                    reviewLine.AddToken(ReviewToken.CreatePunctuationToken(SyntaxKind.ColonToken));
                     first = false;
                 }
-
-                DisplayName(builder, typeInterface);
-            }
-
-            if (!first)
-            {
-                builder.Space();
+                DisplayName(reviewLine, typeInterface);
+                reviewLine.Tokens.Last().HasSuffixSpace = true;
             }
         }
 
-        private static void CloseBrace(CodeFileTokensBuilder builder)
+        private void BuildMember(List<ReviewLine> reviewLines, ISymbol member, bool inHiddenScope)
         {
-            builder.DecrementIndent();
-            builder.WriteIndent();
-            builder.Punctuation(SyntaxKind.CloseBraceToken);
-            builder.NewLine();
-        }
-
-        private void BuildMember(CodeFileTokensBuilder builder, ISymbol member, bool inHiddenScope)
-        {
-            bool isHidden = IsHiddenFromIntellisense(member);
-
-            if (isHidden && !inHiddenScope)
+            bool isHidden = IsHiddenFromIntellisense(member) || inHiddenScope;
+            string lineId = GetLineId(member);
+            var reviewLine = new ReviewLine()
             {
-                builder.Append(null, CodeFileTokenKind.HiddenApiRangeStart);
+                LineId = lineId,
+                IsHidden = isHidden
+            };
+
+            BuildDocumentation(reviewLines, member, isHidden, lineId);
+            BuildAttributes(reviewLines, member.GetAttributes(), isHidden, lineId);
+            reviewLines.Add(reviewLine);
+            DisplayName(reviewLine, member);
+            reviewLine.Tokens.Last().HasSuffixSpace = false;
+
+            // Set member sub kind class for render class styling
+            var memToken = reviewLine.Tokens.FirstOrDefault(m => m.Kind == TokenKind.MemberName);
+            if (memToken != null)
+            {
+                memToken.RenderClasses.Add(member.Kind.ToString().ToLowerInvariant());
             }
-
-            BuildDocumentation(builder, member);
-            BuildAttributes(builder, member.GetAttributes());
-
-            builder.WriteIndent();
-            NodeFromSymbol(builder, member);
 
             if (member.Kind == SymbolKind.Field && member.ContainingType.TypeKind == TypeKind.Enum)
             {
-                builder.Punctuation(SyntaxKind.CommaToken);
+                reviewLine.AddToken(ReviewToken.CreatePunctuationToken(SyntaxKind.CommaToken));
             }
             else if (member.Kind != SymbolKind.Property)
             {
-                builder.Punctuation(SyntaxKind.SemicolonToken);
-            }
-
-            builder.NewLine();
-            if (isHidden && !inHiddenScope)
-            {
-                builder.Append(null, CodeFileTokenKind.HiddenApiRangeEnd);
+                reviewLine.AddToken(ReviewToken.CreatePunctuationToken(SyntaxKind.SemicolonToken));
             }
         }
 
-        private void BuildAttributes(CodeFileTokensBuilder builder, ImmutableArray<AttributeData> attributes)
+        private string GetTypedConstantValue(TypedConstant typedConstant)
+        {
+            return typedConstant.Kind switch
+            {
+                TypedConstantKind.Array => string.Join(", ", typedConstant.Values.Select(v => v.ToString())),
+                _ => typedConstant.Value?.ToString() ?? string.Empty
+            };
+        }
+
+        private void BuildAttributes(List<ReviewLine> reviewLines, ImmutableArray<AttributeData> attributes, bool isHidden, string relatedTo)
         {
             const string attributeSuffix = "Attribute";
             foreach (var attribute in attributes)
             {
-                if ((!IsAccessible(attribute.AttributeClass) &&
-                    attribute.AttributeClass.Name != "FriendAttribute" && 
-                    attribute.AttributeClass.ContainingNamespace.ToString() != "System.Diagnostics.CodeAnalysis")
-                    || IsSkippedAttribute(attribute.AttributeClass))
+                if (attribute.AttributeClass != null)
                 {
-                    continue;
-                }
-                builder.WriteIndent();
-                if(attribute.AttributeClass.DeclaredAccessibility == Accessibility.Internal || attribute.AttributeClass.DeclaredAccessibility == Accessibility.Friend)
-                {
-                    builder.Keyword("internal");
-                    builder.Space();
-                }
-                builder.Punctuation(SyntaxKind.OpenBracketToken);
-                var name = attribute.AttributeClass.Name;
-                if (name.EndsWith(attributeSuffix))
-                {
-                    name = name.Substring(0, name.Length - attributeSuffix.Length);
-                }
-                builder.Append(name, CodeFileTokenKind.TypeName);
-                if (attribute.ConstructorArguments.Any())
-                {
-                    builder.Punctuation(SyntaxKind.OpenParenToken);
-                    bool first = true;
-
-                    foreach (var argument in attribute.ConstructorArguments)
+                    if ((!IsAccessible(attribute.AttributeClass) &&
+                        attribute.AttributeClass.Name != "FriendAttribute" &&
+                        attribute.AttributeClass.ContainingNamespace.ToString() != "System.Diagnostics.CodeAnalysis")
+                        || IsSkippedAttribute(attribute.AttributeClass))
                     {
-                        if (!first)
-                        {
-                            builder.Punctuation(SyntaxKind.CommaToken);
-                            builder.Space();
-                        }
-                        else
-                        {
-                            first = false;
-                        }
-                        BuildTypedConstant(builder, argument);
+                        continue;
                     }
 
-                    foreach (var argument in attribute.NamedArguments)
+                    string args = String.Join(", ", attribute.ConstructorArguments.Select(a => GetTypedConstantValue(a)));
+
+                    var attributeLine = new ReviewLine()
                     {
-                        if (!first)
-                        {
-                            builder.Punctuation(SyntaxKind.CommaToken);
-                            builder.Space();
-                        }
-                        else
-                        {
-                            first = false;
-                        }
-                        builder.Append(argument.Key, CodeFileTokenKind.Text);
-                        builder.Space();
-                        builder.Punctuation(SyntaxKind.EqualsToken);
-                        builder.Space();
-                        BuildTypedConstant(builder, argument.Value);
+                        // GetId() is not unique for attribute class. for e.g. attribute class id is something like "System.FlagsAttribute"
+                        // So, using a unique id for attribute line
+                        LineId = $"{attribute.AttributeClass.GetId()}({args}).{relatedTo}",
+                        IsHidden = isHidden
+                    };
+
+                    if (attribute.AttributeClass.DeclaredAccessibility == Accessibility.Internal || attribute.AttributeClass.DeclaredAccessibility == Accessibility.Friend)
+                    {
+                        attributeLine.AddToken(ReviewToken.CreateKeywordToken("internal"));
                     }
 
-                    builder.Punctuation(SyntaxKind.CloseParenToken);
+                    attributeLine.AddToken(ReviewToken.CreatePunctuationToken(SyntaxKind.OpenBracketToken, false));
+                    var name = attribute.AttributeClass.Name;
+                    if (name.EndsWith(attributeSuffix))
+                    {
+                        name = name.Substring(0, name.Length - attributeSuffix.Length);
+                    }
+                    attributeLine.AddToken(ReviewToken.CreateTypeNameToken(name, false));
+                    if (attribute.ConstructorArguments.Any())
+                    {
+                        attributeLine.AddToken(ReviewToken.CreatePunctuationToken(SyntaxKind.OpenParenToken, false));
+                        bool first = true;
+
+                        foreach (var argument in attribute.ConstructorArguments)
+                        {
+                            if (!first)
+                            {
+                                attributeLine.AddToken(ReviewToken.CreatePunctuationToken(SyntaxKind.CommaToken));
+                            }
+                            else
+                            {
+                                first = false;
+                            }
+                            BuildTypedConstant(attributeLine, argument);
+                        }
+
+                        foreach (var argument in attribute.NamedArguments)
+                        {
+                            if (!first)
+                            {
+                                attributeLine.AddToken(ReviewToken.CreatePunctuationToken(SyntaxKind.CommaToken));
+                            }
+                            else
+                            {
+                                first = false;
+                            }
+                            attributeLine.AddToken(ReviewToken.CreateTextToken(argument.Key));
+                            attributeLine.AddToken(ReviewToken.CreatePunctuationToken(SyntaxKind.EqualsToken));
+                            BuildTypedConstant(attributeLine, argument.Value);
+                        }
+                        attributeLine.Tokens.Last().HasSuffixSpace = false;
+                        attributeLine.AddToken(ReviewToken.CreatePunctuationToken(SyntaxKind.CloseParenToken));
+                    }
+                    attributeLine.Tokens.Last().HasSuffixSpace = false;
+                    attributeLine.AddToken(ReviewToken.CreatePunctuationToken(SyntaxKind.CloseBracketToken));
+                    attributeLine.RelatedToLine = relatedTo;
+                    //Add current attribute line to review lines
+                    reviewLines.Add(attributeLine);
                 }
-                builder.Punctuation(SyntaxKind.CloseBracketToken);
-                builder.NewLine();
             }
+        }
+
+        private bool IsExtensionMemberContainer(INamedTypeSymbol namedType)
+        {
+            if (namedType == null)
+            {
+                return false;
+            }
+                
+            // Extension member containers are compiler-generated nested classes
+            // They have specific name patterns like <G>$ or <M>$ and contain extension marker methods
+
+            // Check if the name matches the compiler-generated pattern for extension containers
+            if (!(namedType.Name.StartsWith("<G>$") || namedType.Name.StartsWith("<M>$")))
+                return false;
+
+            // Check if this type contains extension marker methods or nested types with extension markers
+            return namedType.GetMembers().OfType<IMethodSymbol>()
+                .Any(m => IsExtensionMarkerMethod(m)) ||
+                namedType.GetTypeMembers().Any(t => t.GetMembers().OfType<IMethodSymbol>()
+                    .Any(m => IsExtensionMarkerMethod(m)));
+        }
+
+        private bool IsExtensionMarkerMethod(IMethodSymbol method)
+        {
+            if (method == null)
+                return false;
+
+            // Extension marker methods are compiler-generated methods with specific characteristics:
+            // - They are named exactly "<Extension>$" 
+            // - They have void return type and take the extension target as parameter
+            // Note: We don't check for CompilerGeneratedAttribute as it may not be accessible in the symbol model
+            return method.Name == "<Extension>$" &&
+                   method.ReturnType?.SpecialType == SpecialType.System_Void &&
+                   method.Parameters.Length >= 1;
+        }
+
+        private void BuildExtensionMember(List<ReviewLine> reviewLines, INamedTypeSymbol extensionContainer, bool inHiddenScope)
+        {
+            if (extensionContainer == null)
+            {
+                return;
+            }
+
+            bool isHidden = IsHiddenFromIntellisense(extensionContainer) || inHiddenScope;
+
+            // Find the extension marker method to get the extension target type
+            // Check both direct methods and methods in nested types
+            var extensionMarker = extensionContainer.GetMembers().OfType<IMethodSymbol>()
+                .FirstOrDefault(m => IsExtensionMarkerMethod(m));
+
+            if (extensionMarker == null)
+            {
+                // Check nested types for extension markers
+                foreach (var nestedType in extensionContainer.GetTypeMembers())
+                {
+                    extensionMarker = nestedType.GetMembers().OfType<IMethodSymbol>()
+                        .FirstOrDefault(m => IsExtensionMarkerMethod(m));
+                    if (extensionMarker != null)
+                    {
+                        break;
+                    }
+                }
+            }
+
+            if (extensionMarker == null || extensionMarker.Parameters.Length == 0)
+            {
+                return;
+            }
+
+            var targetType = extensionMarker.Parameters[0].Type;
+            if (targetType == null)
+            {
+                return;
+            }
+
+            // Create the extension declaration line
+            var extensionLine = new ReviewLine()
+            {
+                LineId = extensionContainer.GetId(),
+                IsHidden = isHidden
+            };
+
+            extensionLine.Tokens.Add(ReviewToken.CreateKeywordToken("extension"));
+            extensionLine.Tokens.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.OpenParenToken, false));
+            DisplayName(extensionLine, targetType);
+            extensionLine.Tokens.Last().HasSuffixSpace = true;
+
+            // Add parameter name (usually the target type name in lowercase)
+            var parameterName = extensionMarker.Parameters[0].Name;
+            extensionLine.Tokens.Add(ReviewToken.CreateTextToken(parameterName, hasSuffixSpace: false));
+            extensionLine.Tokens.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.CloseParenToken));
+            extensionLine.Tokens.Last().HasSuffixSpace = true;
+            extensionLine.Tokens.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.OpenBraceToken));
+
+            // Add all non-extension-marker members to the extension block
+            // Include members from both the container and any nested types
+            var membersToProcess = new List<ISymbol>();
+            membersToProcess.AddRange(extensionContainer.GetMembers().Where(m =>
+                m.Kind != SymbolKind.NamedType && !m.IsImplicitlyDeclared && IsAccessible(m)));
+
+            foreach (var nestedType in extensionContainer.GetTypeMembers())
+            {
+                membersToProcess.AddRange(nestedType.GetMembers().Where(m =>
+                    m.Kind != SymbolKind.NamedType && !m.IsImplicitlyDeclared && IsAccessible(m)));
+            }
+
+            foreach (var member in SymbolOrderProvider.OrderMembers(membersToProcess))
+            {
+                // Skip extension marker methods
+                if (member is IMethodSymbol method && IsExtensionMarkerMethod(method)) continue;
+
+                // Skip property/event accessors
+                if (member is IMethodSymbol m)
+                {
+                    if (m.MethodKind == MethodKind.PropertyGet ||
+                        m.MethodKind == MethodKind.PropertySet ||
+                        m.MethodKind == MethodKind.EventAdd ||
+                        m.MethodKind == MethodKind.EventRemove ||
+                        m.MethodKind == MethodKind.EventRaise)
+                    {
+                        continue;
+                    }
+                }
+                BuildMember(extensionLine.Children, member, isHidden);
+            }
+
+            reviewLines.Add(extensionLine);
+            reviewLines.Add(new ReviewLine()
+            {
+                Tokens = [
+                    ReviewToken.CreatePunctuationToken(SyntaxKind.CloseBraceToken)
+                ],
+                IsHidden = isHidden,
+                IsContextEndLine = true
+            });
+            reviewLines.Add(new ReviewLine() { IsHidden = isHidden, RelatedToLine = extensionLine.LineId });
         }
 
         private bool IsSkippedAttribute(INamedTypeSymbol attributeAttributeClass)
@@ -536,6 +714,9 @@ namespace ApiView
                 case "EditorBrowsableAttribute":
                 case "NullableAttribute":
                 case "NullableContextAttribute":
+                case "RequiresUnreferencedCodeAttribute":
+                case "RequiresDynamicCodeAttribute":
+                case "CompilerGeneratedAttribute":
                     return true;
                 default:
                     return false;
@@ -544,10 +725,10 @@ namespace ApiView
 
         private bool IsHiddenFromIntellisense(ISymbol member) =>
             member.GetAttributes().Any(d => d.AttributeClass?.Name == "EditorBrowsableAttribute"
-                                            && (EditorBrowsableState) d.ConstructorArguments[0].Value == EditorBrowsableState.Never)
-            || CodeFileBuilder.IsExplicitInterfaceImplementation(member);
+                                            && (EditorBrowsableState)d.ConstructorArguments[0].Value! == EditorBrowsableState.Never)
+            || IsExplicitInterfaceImplementation(member);
 
-        private static bool IsExplicitInterfaceImplementation(ISymbol symbol)
+        private bool IsExplicitInterfaceImplementation(ISymbol symbol)
         {
             return symbol switch
             {
@@ -560,27 +741,31 @@ namespace ApiView
         private bool IsDecoratedWithAttribute(ISymbol member, string attributeName) =>
             member.GetAttributes().Any(d => d.AttributeClass?.Name == attributeName);
 
-        private void BuildTypedConstant(CodeFileTokensBuilder builder, TypedConstant typedConstant)
+        private void BuildTypedConstant(ReviewLine reviewLine, TypedConstant typedConstant)
         {
+            var tokenList = reviewLine.Tokens;
             if (typedConstant.IsNull)
             {
-                builder.Keyword(SyntaxKind.NullKeyword);
+                tokenList.Add(ReviewToken.CreateKeywordToken(SyntaxKind.NullKeyword, false));
             }
             else if (typedConstant.Kind == TypedConstantKind.Enum)
             {
-                new CodeFileBuilderEnumFormatter(builder).Format(typedConstant.Type, typedConstant.Value);
+                new CodeFileBuilderEnumFormatter(tokenList).Format(typedConstant.Type, typedConstant.Value);
             }
             else if (typedConstant.Kind == TypedConstantKind.Type)
             {
-                builder.Keyword(SyntaxKind.TypeOfKeyword);
-                builder.Punctuation("(");
-                DisplayName(builder, (ITypeSymbol)typedConstant.Value);
-                builder.Punctuation(")");
+                tokenList.Add(ReviewToken.CreateKeywordToken(SyntaxKind.TypeOfKeyword, false));
+                tokenList.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.OpenParenToken, false));
+                DisplayName(reviewLine, (ITypeSymbol)typedConstant.Value!);
+                reviewLine.Tokens.Last().HasSuffixSpace = false;
+                tokenList.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.CloseParenToken, false));
             }
             else if (typedConstant.Kind == TypedConstantKind.Array)
             {
-                builder.Keyword(SyntaxKind.NewKeyword);
-                builder.Punctuation("[] {");
+                tokenList.Add(ReviewToken.CreateKeywordToken(SyntaxKind.NewKeyword));
+                tokenList.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.OpenBracketToken, false));
+                tokenList.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.CloseBracketToken));
+                tokenList.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.OpenBraceToken));
 
                 bool first = true;
 
@@ -588,59 +773,47 @@ namespace ApiView
                 {
                     if (!first)
                     {
-                        builder.Punctuation(SyntaxKind.CommaToken);
-                        builder.Space();
+                        tokenList.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.CommaToken));
                     }
                     else
                     {
                         first = false;
                     }
 
-                    BuildTypedConstant(builder, value);
+                    BuildTypedConstant(reviewLine, value);
+                    reviewLine.Tokens.Last().HasSuffixSpace = false;
                 }
-                builder.Punctuation("}");
+                tokenList.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.CloseBraceToken, false));
             }
             else
             {
                 if (typedConstant.Value is string s)
                 {
-                    builder.Append(
-                        ObjectDisplay.FormatLiteral(s, ObjectDisplayOptions.UseQuotes | ObjectDisplayOptions.EscapeNonPrintableCharacters),
-                        CodeFileTokenKind.StringLiteral);
+                    tokenList.Add(ReviewToken.CreateStringLiteralToken(ObjectDisplay.FormatLiteral(s, ObjectDisplayOptions.UseQuotes | ObjectDisplayOptions.EscapeNonPrintableCharacters), false));
                 }
                 else
                 {
-                    builder.Append(
-                        ObjectDisplay.FormatPrimitive(typedConstant.Value, ObjectDisplayOptions.None),
-                        CodeFileTokenKind.Literal);
+                    tokenList.Add(ReviewToken.CreateLiteralToken(ObjectDisplay.FormatPrimitive(typedConstant.Value, ObjectDisplayOptions.None), false));
                 }
             }
         }
 
-        private void NodeFromSymbol(CodeFileTokensBuilder builder, ISymbol symbol)
+        private void BuildVisibility(List<ReviewToken> tokenList, ISymbol symbol)
         {
-            builder.Append(new CodeFileToken()
-            {
-                DefinitionId = symbol.GetId(),
-                Kind = CodeFileTokenKind.LineIdMarker
-            });
-            DisplayName(builder, symbol, symbol);
+            tokenList.Add(ReviewToken.CreateKeywordToken(ToEffectiveAccessibility(symbol.DeclaredAccessibility)));
         }
 
-        private void BuildVisibility(CodeFileTokensBuilder builder, ISymbol symbol)
+        private void DisplayName(ReviewLine reviewLine, ISymbol symbol, ISymbol definedSymbol = null)
         {
-            builder.Keyword(SyntaxFacts.GetText(ToEffectiveAccessibility(symbol.DeclaredAccessibility)));
-        }
+            var reviewLineTokens = reviewLine.Tokens;
 
-        private void DisplayName(CodeFileTokensBuilder builder, ISymbol symbol, ISymbol definedSymbol = null)
-        {
             if (NeedsAccessibility(symbol))
             {
-                builder.Keyword(SyntaxFacts.GetText(ToEffectiveAccessibility(symbol.DeclaredAccessibility)));
-                builder.Space();
+                reviewLineTokens.Add(ReviewToken.CreateKeywordToken(ToEffectiveAccessibility(symbol.DeclaredAccessibility)));
             }
             if (symbol is IPropertySymbol propSymbol && propSymbol.DeclaredAccessibility != Accessibility.Internal)
             {
+                SymbolDisplayPart previous = default(SymbolDisplayPart);
                 var parts = propSymbol.ToDisplayParts(_defaultDisplayFormat);
                 for (int i = 0; i < parts.Length; i++)
                 {
@@ -652,14 +825,39 @@ namespace ApiView
                             i++;
                         }
                     }
-                    builder.Append(MapToken(definedSymbol, parts[i]));
+                    var previousToken = reviewLine.Tokens.LastOrDefault();
+                    //Add a new code line as child if there is a line break
+                    if (parts[i].Kind == SymbolDisplayPartKind.LineBreak)
+                    {
+                        var subLine = new ReviewLine()
+                        {
+                            LineId = definedSymbol.GetId(),
+                        };
+                        reviewLine.Children.Add(subLine);
+                        reviewLineTokens = subLine.Tokens;
+                    }
+                    var token = MapToken(definedSymbol: definedSymbol!, symbolDisplayPart: parts[i],
+                        previousSymbolDisplayPart: previous, previousToken);
+                    if (token != null)
+                    {
+                        reviewLineTokens.Add(token);
+                    }
+                    previous = parts[i];
                 }
             }
             else
             {
+                SymbolDisplayPart previous = default(SymbolDisplayPart);
                 foreach (var symbolDisplayPart in symbol.ToDisplayParts(_defaultDisplayFormat))
                 {
-                    builder.Append(MapToken(definedSymbol, symbolDisplayPart));
+                    var previousToken = reviewLine.Tokens.LastOrDefault();
+                    var token = MapToken(definedSymbol: definedSymbol!, symbolDisplayPart: symbolDisplayPart,
+                        previousSymbolDisplayPart: previous, previousToken: previousToken);
+                    if (token != null)
+                    {
+                        reviewLineTokens.Add(token);
+                    }
+                    previous = symbolDisplayPart;
                 }
             }
         }
@@ -679,9 +877,21 @@ namespace ApiView
             };
         }
 
-        private CodeFileToken MapToken(ISymbol definedSymbol, SymbolDisplayPart symbolDisplayPart)
+        private ReviewToken MapToken(ISymbol definedSymbol, SymbolDisplayPart symbolDisplayPart, SymbolDisplayPart previousSymbolDisplayPart, ReviewToken previousToken)
         {
-            CodeFileTokenKind kind;
+            string navigateToId = null;
+            var symbol = symbolDisplayPart.Symbol;
+
+            if (symbol is INamedTypeSymbol &&
+                (definedSymbol == null || !SymbolEqualityComparer.Default.Equals(definedSymbol, symbol)) &&
+                SymbolEqualityComparer.Default.Equals(_assembly, symbol.ContainingAssembly))
+            {
+                navigateToId = symbol.GetId();
+            }
+
+            var tokenValue = symbolDisplayPart.ToString();
+
+            ReviewToken token = null;
 
             switch (symbolDisplayPart.Kind)
             {
@@ -694,22 +904,22 @@ namespace ApiView
                 case SymbolDisplayPartKind.ErrorTypeName:
                 case SymbolDisplayPartKind.InterfaceName:
                 case SymbolDisplayPartKind.StructName:
-                    kind = CodeFileTokenKind.TypeName;
+                    token = ReviewToken.CreateTypeNameToken(tokenValue, false);
                     break;
                 case SymbolDisplayPartKind.Keyword:
-                    kind = CodeFileTokenKind.Keyword;
-                    break;
-                case SymbolDisplayPartKind.LineBreak:
-                    kind = CodeFileTokenKind.Newline;
+                    token = ReviewToken.CreateKeywordToken(tokenValue, false);
                     break;
                 case SymbolDisplayPartKind.StringLiteral:
-                    kind = CodeFileTokenKind.StringLiteral;
+                    token = ReviewToken.CreateStringLiteralToken(tokenValue, false);
                     break;
                 case SymbolDisplayPartKind.Punctuation:
-                    kind = CodeFileTokenKind.Punctuation;
+                    token = ReviewToken.CreatePunctuationToken(tokenValue, false);
                     break;
                 case SymbolDisplayPartKind.Space:
-                    kind = CodeFileTokenKind.Whitespace;
+                    if (previousToken != null)
+                    {
+                        previousToken.HasSuffixSpace = true;
+                    }
                     break;
                 case SymbolDisplayPartKind.PropertyName:
                 case SymbolDisplayPartKind.EventName:
@@ -719,30 +929,17 @@ namespace ApiView
                 case SymbolDisplayPartKind.EnumMemberName:
                 case SymbolDisplayPartKind.ExtensionMethodName:
                 case SymbolDisplayPartKind.ConstantName:
-                    kind = CodeFileTokenKind.MemberName;
+                    token = ReviewToken.CreateMemberNameToken(tokenValue, false);
                     break;
                 default:
-                    kind = CodeFileTokenKind.Text;
+                    token = ReviewToken.CreateTextToken(tokenValue, hasSuffixSpace: false);
                     break;
             }
-
-            string navigateToId = null;
-            var symbol = symbolDisplayPart.Symbol;
-
-            if (symbol is INamedTypeSymbol &&
-                (definedSymbol == null || !SymbolEqualityComparer.Default.Equals(definedSymbol, symbol)) &&
-                SymbolEqualityComparer.Default.Equals(_assembly, symbol.ContainingAssembly))
+            if (token != null && !String.IsNullOrWhiteSpace(navigateToId))
             {
-                navigateToId = symbol.GetId();
+                token.NavigateToId = navigateToId!;
             }
-
-            return new CodeFileToken()
-            {
-                DefinitionId = (definedSymbol != null && SymbolEqualityComparer.Default.Equals(definedSymbol, symbol)) ? definedSymbol.GetId() : null,
-                NavigateToId = navigateToId,
-                Value = symbolDisplayPart.ToString(),
-                Kind = kind
-            };
+            return token;
         }
 
         private Accessibility ToEffectiveAccessibility(Accessibility accessibility)
@@ -767,7 +964,7 @@ namespace ApiView
                 case Accessibility.Public:
                     return true;
                 case Accessibility.Internal:
-                    return s.GetAttributes().Any(a => a.AttributeClass.Name == "FriendAttribute");
+                    return s.GetAttributes().Any(a => a.AttributeClass?.Name == "FriendAttribute");
                 default:
                     return IsAccessibleExplicitInterfaceImplementation(s);
             }
@@ -785,11 +982,11 @@ namespace ApiView
 
         internal class CodeFileBuilderEnumFormatter : AbstractSymbolDisplayVisitor
         {
-            private readonly CodeFileTokensBuilder _builder;
+            private readonly List<ReviewToken> _tokenList;
 
-            public CodeFileBuilderEnumFormatter(CodeFileTokensBuilder builder) : base(null, SymbolDisplayFormat.FullyQualifiedFormat, false, null, 0, false)
+            public CodeFileBuilderEnumFormatter(List<ReviewToken> tokenList) : base(null, SymbolDisplayFormat.FullyQualifiedFormat, false, null, 0, false)
             {
-                _builder = builder;
+                _tokenList = tokenList;
             }
 
             protected override AbstractSymbolDisplayVisitor MakeNotFirstVisitor(bool inNamespaceOrType = false)
@@ -799,35 +996,108 @@ namespace ApiView
 
             protected override void AddLiteralValue(SpecialType type, object value)
             {
-                _builder.Append(ObjectDisplay.FormatPrimitive(value, ObjectDisplayOptions.None), CodeFileTokenKind.Literal);
+                _tokenList.Add(ReviewToken.CreateLiteralToken(ObjectDisplay.FormatPrimitive(value, ObjectDisplayOptions.None)));
             }
 
             protected override void AddExplicitlyCastedLiteralValue(INamedTypeSymbol namedType, SpecialType type, object value)
             {
-                _builder.Append(ObjectDisplay.FormatPrimitive(value, ObjectDisplayOptions.None), CodeFileTokenKind.Literal);
+                _tokenList.Add(ReviewToken.CreateLiteralToken(ObjectDisplay.FormatPrimitive(value, ObjectDisplayOptions.None)));
             }
 
             protected override void AddSpace()
             {
-                _builder.Space();
+                var lastToken = _tokenList.LastOrDefault();
+                if (lastToken != null)
+                {
+                    lastToken.HasSuffixSpace = true;
+                }
             }
 
             protected override void AddBitwiseOr()
             {
-                _builder.Punctuation(SyntaxKind.BarToken);
+                if (_tokenList.Count > 0)
+                    _tokenList.Last().HasSuffixSpace = true;
+                _tokenList.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.BarToken));
             }
 
             public override void VisitField(IFieldSymbol symbol)
             {
-                _builder.Append(symbol.Type.Name, CodeFileTokenKind.TypeName);
-                _builder.Punctuation(SyntaxKind.DotToken);
-                _builder.Append(symbol.Name, CodeFileTokenKind.MemberName);
+                _tokenList.Add(ReviewToken.CreateTypeNameToken(symbol.Type.Name, false));
+                _tokenList.Add(ReviewToken.CreatePunctuationToken(SyntaxKind.DotToken, false));
+                _tokenList.Add(ReviewToken.CreateMemberNameToken(symbol.Name, false));
             }
 
             public void Format(ITypeSymbol type, object typedConstantValue)
             {
                 AddNonNullConstantValue(type, typedConstantValue);
             }
+        }
+
+        string GetLineId(ISymbol member)
+        {
+            string lineId = member.GetId();
+            var value = (IsExplicitInterfaceImplementation(member), member) switch
+            {
+                (true, IMethodSymbol method) => $"{method.ExplicitInterfaceImplementations.First().ContainingType.GetId()}.{lineId}",
+                (true, IPropertySymbol prop) => $"{prop.ExplicitInterfaceImplementations.First().ContainingType.GetId()}.{lineId}",
+                (_, _) => lineId
+            };
+            return value;
+        }
+
+        /// <summary>
+        /// Verifies that all LineId values in the codeFile's ReviewLines collection are unique.
+        /// </summary>
+        /// <param name="codeFile">The CodeFile to verify.</param>
+        /// <exception cref="InvalidOperationException">Thrown when duplicate LineId values are found.</exception>
+        private static void VerifyCodeFile(CodeFile codeFile)
+        {
+            var lineIds = new HashSet<string>();
+            var duplicates = new List<string>();
+
+            void CheckLineIds(IEnumerable<ReviewLine> lines)
+            {
+                foreach (var line in lines)
+                {
+                    if (!string.IsNullOrEmpty(line.LineId))
+                    {
+                        if (!lineIds.Add(line.LineId))
+                        {
+                            duplicates.Add(line.LineId);
+                        }
+                    }
+
+                    if (line.Children != null && line.Children.Count > 0)
+                    {
+                        CheckLineIds(line.Children);
+                    }
+                }
+            }
+
+            CheckLineIds(codeFile.ReviewLines);
+
+            if (duplicates.Count > 0)
+            {
+                throw new InvalidOperationException(
+                    $"Duplicate LineId values found: {string.Join(", ", duplicates.Distinct())}");
+            }
+        }
+
+        public static string GetPackageVersion(IAssemblySymbol assemblySymbol)
+        {
+            var attr = assemblySymbol
+                .GetAttributes()
+                .FirstOrDefault(a => a.AttributeClass?.Name == "AssemblyInformationalVersionAttribute");
+
+            if (attr == null || attr.ConstructorArguments.Length == 0)
+                return null;
+            
+            var infoVersion = attr.ConstructorArguments[0].Value as string;
+            if (string.IsNullOrWhiteSpace(infoVersion))
+                return null;
+
+            var version = infoVersion.Split('+')[0].Trim();
+            return version.Length == 0 ? null : version;
         }
     }
 }
